Chat message saved locally: simplify separation by opponent
It seems surprising than the opponent info was not sent already. If we think it's worth enough to send info can be kept that way, otherwise an idea would be to get the info from `Line`.

Also considering the message is saved in temp storage I don't think it's worth troubling ourselves with deletion.